[IMP]添加禁用crontab选项

+++ docker-entrypoint.sh
@@ -1,11 +1,15 @@
 #!/bin/sh
-REFRESH_TOKEN=${REFRESH_TOKEN:-"0 * * * *"}
-REFRESH_CACHE=${REFRESH_CACHE:-"*/10 * * * *"}
+
+if [ -z $DISABLE_CRON ];then
+    REFRESH_TOKEN=${REFRESH_TOKEN:-"0 * * * *"}
+    REFRESH_CACHE=${REFRESH_CACHE:-"*/10 * * * *"}
+    rm -rf /tmp/cron.`whoami`
+    echo "${REFRESH_TOKEN} php /var/www/html/one.php token:refresh" >> /tmp/cron.`whoami`
+    echo "${REFRESH_CACHE} php /var/www/html/one.php cache:refresh" >> /tmp/cron.`whoami`
+    crontab -u `whoami` /tmp/cron.`whoami`
+    crond
+fi
+
 chown -R www-data:www-data /var/www/html/cache
 chown -R www-data:www-data /var/www/html/config
-rm -rf /tmp/cron.`whoami`
-echo "${REFRESH_TOKEN} php /var/www/html/one.php token:refresh" >> /tmp/cron.`whoami`
-echo "${REFRESH_CACHE} php /var/www/html/one.php cache:refresh" >> /tmp/cron.`whoami`
-crontab -u `whoami` /tmp/cron.`whoami`
-crond
 php-fpm & nginx '-g daemon off;'
